Crawl daily BAR data:
	1) Update the fetcu_url return value.

Submitted by Guixian Lin

+++ src/data/receive/bar/daily/crawl_daily.h
@@ -12,7 +12,7 @@
 int initialize(void);
 void destroy(void);
 
-int fetch_url(const *url, uint8_t *data, size_t len);
+size_t fetch_url(const *url, uint8_t *data, size_t len);
 
 int save(uint8_t *data, size_t size, const char *dst_fn);
